llama-deploy-testCase/pyproject.toml
[tool.poetry]
name = "llama-agent-testcase"
version = "0.1.0"
description = ""
authors = ["Bless50 <115646790+Bless50@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8.1,<3.12"
llama-index = "^0.11.6"
llama-deploy = "^0.1.0"
llama-index-core = "^0.11.6"
llama-index-utils-workflow = "^0.2.1"
python-dotenv = "^1.0.1"
llama-index-llms-groq = "^0.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
